+++ backend/views/file-storage/_form.php
@@ -1,11 +1,11 @@
 <?php
 
 use yii\helpers\Html;
-use yii\widgets\ActiveForm;
+use yii\bootstrap\ActiveForm;
 
 /* @var $this yii\web\View */
-/* @var $model common\models\FileStorageItem */
-/* @var $form yii\widgets\ActiveForm */
+/* @var $model trntv\filekit\storage\models\FileStorageItem */
+/* @var $form yii\bootstrap\ActiveForm */
 ?>
 
 <div class="file-storage-item-form">