+++ backend/views/log/_search.php
@@ -1,11 +1,11 @@
 <?php
 
 use yii\helpers\Html;
-use yii\widgets\ActiveForm;
+use yii\bootstrap\ActiveForm;
 
 /* @var $this yii\web\View */
-/* @var $model backend\\models\search\SystemLogSearch */
-/* @var $form yii\widgets\ActiveForm */
+/* @var $model backend\models\search\SystemLogSearch */
+/* @var $form yii\bootstrap\ActiveForm */
 ?>
 
 <div class="system-log-search">
@@ -25,7 +25,7 @@ use yii\widgets\ActiveForm;
 
     <?= $form->field($model, 'prefix') ?>
 
-    <?php // echo $form->field($model, 'message') ?>
+    <?= $form->field($model, 'message') ?>
 
     <div class="form-group">
         <?= Html::submitButton(Yii::t('backend', 'Search'), ['class' => 'btn btn-primary']) ?>