+++ backend/views/page/_search.php
@@ -1,11 +1,11 @@
 <?php
 
 use yii\helpers\Html;
-use yii\widgets\ActiveForm;
+use yii\bootstrap\ActiveForm;
 
 /* @var $this yii\web\View */
-/* @var $model backend\\models\search\PageSearch */
-/* @var $form yii\widgets\ActiveForm */
+/* @var $model backend\models\search\PageSearch */
+/* @var $form yii\bootstrap\ActiveForm */
 ?>
 
 <div class="page-search">