+++ backend/views/user/_search.php
@@ -1,11 +1,11 @@
 <?php
 
 use yii\helpers\Html;
-use yii\widgets\ActiveForm;
+use yii\bootstrap\ActiveForm;
 
 /* @var $this yii\web\View */
-/* @var $model backend\\models\search\UserSearch */
-/* @var $form yii\widgets\ActiveForm */
+/* @var $model backend\models\search\UserSearch */
+/* @var $form yii\bootstrap\ActiveForm */
 ?>
 
 <div class="user-search">
@@ -21,19 +21,15 @@ use yii\widgets\ActiveForm;
 
     <?= $form->field($model, 'auth_key') ?>
 
-    <?= $form->field($model, 'password_hash') ?>
+    <?php echo $form->field($model, 'email') ?>
 
-    <?= $form->field($model, 'password_reset_token') ?>
+    <?php echo $form->field($model, 'role') ?>
 
-    <?php // echo $form->field($model, 'email') ?>
+    <?php echo $form->field($model, 'status') ?>
 
-    <?php // echo $form->field($model, 'role') ?>
+    <?php echo $form->field($model, 'created_at') ?>
 
-    <?php // echo $form->field($model, 'status') ?>
-
-    <?php // echo $form->field($model, 'created_at') ?>
-
-    <?php // echo $form->field($model, 'updated_at') ?>
+    <?php echo $form->field($model, 'updated_at') ?>
 
     <div class="form-group">
         <?= Html::submitButton(Yii::t('backend', 'Search'), ['class' => 'btn btn-primary']) ?>