+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Yii Framework 2 Change Log
 
 1.4.0 under development
 -----------------------
+- Enh: Backend now use `yii\bootstrap\ActiveForm` instead of `yii\widgets\ActiveForm`
 - Enh: added `getFullName` for `UserProfile` and `getPublicIdentity` for `User`
 - Fix: added some settings to prevent postfix `fatal: usage: sendmail [options]` error
 - Enh: Gii Module has separate configs for backend and frontend